#140 - Are scope requirements meant to be transitive?
If a scope depends on a parent scope and that parent scope has
external dependencies - then the scope should also be able to
depend on those external dependencies.

Ultimately the fix for this is in ScopeInfo.providesDependency()
adding in the check to see if the dependency is in requires
(is an external dependency of the scope).