#143 - Remove service loader META-INF services file for custom modules
Now that custom modules can have no default constructor they can not be loaded via ServiceLoader. This should not matter as it really didn't make sense to service load the custom modules anyway.

Note that the 'default' scope modules still are expected to be service loaded. That does mean that we do not create a specific constructor for 'default' scope modules. External dependencies for default modules must still be provided via withBean()