#243 - Fix for ClassCastException with generic types
The core fix is with DBeanMap. In detail, when wiring mocks/spy
this part of DBeanMap additionally registers the mock
with the generic type (which is in the generated code).

Prior to this fix, this process was too loose and registered
the mock with other types like Model2 in the test case.