Fix for @Component.Import for partial compile - Add @DependencyMeta importedComponent property
For partial compile the meta data needs to be read from the @DependencyMeta to restore the state and for imported components we need to identify these as there $DI code is generated into a .di package to avoid split package issues with Java modules.

This change uses the explicit @DependencyMeta importedComponent property to identify those imported types (when a partial compile does not include them).