Fix for partial compile failing for external optional Dependency
In the tests this dependency is org.example.myapp.config.SomeOptionalDep. It gets added to autoRequires and in a full compile. For a partial compile this isn't restored from metadata and partial compile fails complaining that the dependency is not provided.

This fix is to mark this dependency as a soft: dependency. Partial compile reads the soft: dependency from metadata.

An alternative would be to add a mechanism (extra metadata annotation to save and restore the autoRequires types.