#555 No qualifier injection point prefers no qualifier bean (over named bean)
In the case where we have 2 beans of the same type and one has no qualifier,
then when injecting a bean with no qualifier specified then prefer to wire
the bean with no qualifier.

Example: In the MegaStoreManager example we have NoNameStore as a bean with
no name qualifier and that is preferred for the noQualifier injection point.