Fix for unnamed generic types returned by @Factory @Bean
As currently the full generic type is not used in the generated build method name but just the top level main type then the generated build method can clash resulting in a compile error.

This fix detects this case and includes the factory method name in the "generated build method" such that it avoids the clash.