Fix partial compile regression introduced in 10.1 via #645 (#654)
* Fix partial compile regression introduced in 10.1 via #645

Fixes a regression that breaks partial compilation when qualifiers are used.

Regression introduced in 10.1 via:
https://github.com/avaje/avaje-inject/pull/645/files

With this regression we observe compile errors stating that dependencies with qualifiers are missing. When reviewing the relevant generated @MetaData dependsOn attributes in the generate source we see that there are qualifiers incorrectly appended to the factory type.

The fix is in the MetaData constructor, such that the dependsOn does NOT use the name/qualifier of the type.