Failing test for ModuleOrdering that skips the extra test module(s) (#760)
* Failing test for ModuleOrdering that skips the extra test module(s)

Test components in src/test that are not annotated with @TestScope go into an extra module. This module is ignored/skipped by the compiled module ordering.

A fix for this is desired.

* Add ModuleOrdering.supportsExpected() such that we only use compiled module ordering when ...

When it supports the modules that are being used. When it does not then fall back to the default ordering mechanism.

* Opps, need to set factoryOrder variable there ...

* [test] Add test injecting BeanScope into a test component

---------

Co-authored-by: Josiah Noel <32279667+SentryMan@users.noreply.github.com>