Bug fix for @Bean returning Optional not registering extended type (#782)
Example:
```
@Bean
Optional<CFace> optional() { ... }
```

The generated isBeanAbsent() method should include all the types that CFace extends such as interfaces. If CFace extends C2Face then both should be there like:

.isBeanAbsent(CFace.class, C2Face.class))

The bug was that only the top type of CFace was included like:

```
.isBeanAbsent(CFace.class))
```

This was due to a bug in MethodReader determining the incorrect returnElement for the optionalType = true case. The returnElement was the Optional type rather than the parameter type CFace.