Graceful shutdown - defaults for io.ebean.Database to shutdown() last
For graceful shutdown, we desire components to shutdown in an appropriate order. For io.ebean.Database it should be shutdown last.

Adds a default shutdown method and shutdown priority for special components such that they default to doing "the right thing" wrt graceful shutdown.

For ebean:

```java
  @Bean
  Database database()
```
We desire that to default to:
```java
  @Bean(destroyMethod = "shutdown", destroyPriority = Integer.MAX_VALUE)
  Database database()
```