Improve parent bean and scope detection
Previously the case in Mod4Scope would fail as the only detected parent was the Mod3Scope (at least, in the way that ParentageTest is building the scopes).

Contains also wasn't checking if the parent existed and contained the bean, so get/getOptional would be able to return it but contains would say it couldn't.

Also added a few Nullable annotations because my IDE was complaining that some of the fields/params could never be null